Fix nova db post setup tasks to remove race
If the nova_db_post_setup.yml task is delegated to the same conductor
host, the tasks run in parallel. If the nova_db_post_setup.yml task is
delegated to different conductor hosts, the tasks run in serial. When
the tasks run in parallel this creates a race condition which can cause
the "Perform a cell_v2 discover" task to fail due to multiple tasks
attempting to make the same changes to the database.

The tasks in nova_db_post_setup.yml should only need to be run once for
the new compute hosts, not for each one. This change limits the tasks to
running once to prevent the possibility of conflicting database
changes being attempted.

Closes-bug: 1752540
(cherry picked from commit 9720b9a087fc1b3ceb1d8dff97d14baae17184f6)